fixed md, added sxcu files
