+++ .gitattributes
@@ -1,3 +1,3 @@
-server/uploads/* linguist-vendored
+src/server/uploads/* linguist-vendored
 .github/* linguist-vendored
-server/views/* linguist-vendored
+src/server/views/* linguist-vendored
