+++ .gitignore
@@ -1,3 +1,3 @@
-/node_modules
-config.real.json
-package-lock.json
+src/node_modules
+src/config.real.json
+src/package-lock.json
