+++ src/bot/commands/deleteFile.js
