+++ src/bot/commands/eval.js
