+++ src/bot/commands/exec.js
