+++ src/bot/commands/help.js
