+++ src/bot/commands/restart.js
