+++ src/bot/events/ready.js
