+++ src/server/routes/gallery.js
