+++ src/server/views/galleryLogin.ejs
