+++ src/server/views/paste.ejs
